+++ CHANGELOG.md
@@ -9,4 +9,5 @@
   - create class object for Base, no need to define tablename
   - no need to import Base from dbjson
   - add multiple data
+  - no need to import data types from sqlalchemy